Voting App AngularJS
Voting App using AngularJS. Allows the user to enter a question and unlimited answers. After creating the poll the user will be able to vote for their answers.

Note: Currently this application is only utilizing the front-end so the app only works for the user accessing the page. 

Technology: HTML, SCSS and AngularJS

Link: http://votingpoll.teammeet.x10host.com/